Split test_cache.py to separate test modules
