+++ tests/cache/test_cache_save_evaluation.py
@@ -9,79 +9,6 @@ from statue.constants import DEFAULT_HISTORY_SIZE
 from statue.exceptions import CacheError
 
 
-def test_cache_constructor_with_none_root_directory(tmp_path):
-    cache = Cache()
-    assert cache.cache_root_directory is None
-    assert cache.evaluations_dir is None
-    assert not cache.all_evaluation_paths
-
-
-def test_cache_constructor_with_non_existing_directory(tmp_path):
-    cache_dir = tmp_path / "cache"
-    assert not cache_dir.exists()
-
-    cache = Cache(cache_dir)
-    assert cache.cache_root_directory == cache_dir
-    assert cache_dir.exists()
-    assert cache.evaluations_dir == cache_dir / "evaluations"
-    assert cache.evaluations_dir.exists()
-    assert not cache.all_evaluation_paths
-
-
-def test_cache_constructor_with_existing_directory(tmp_path):
-    cache_dir = tmp_path / "cache"
-    cache_dir.mkdir()
-
-    cache = Cache(cache_dir)
-
-    assert cache.cache_root_directory == cache_dir
-    assert cache_dir.exists()
-    assert cache.evaluations_dir == cache_dir / "evaluations"
-    assert cache.evaluations_dir.exists()
-    assert not cache.all_evaluation_paths
-
-
-def test_cache_constructor_with_evaluations_directory_already_existing(tmp_path):
-    cache_dir = tmp_path / "cache"
-    (cache_dir / "evaluations").mkdir(parents=True)
-
-    cache = Cache(cache_dir)
-
-    assert cache.cache_root_directory == cache_dir
-    assert cache_dir.exists()
-    assert cache.evaluations_dir == cache_dir / "evaluations"
-    assert cache.evaluations_dir.exists()
-    assert not cache.all_evaluation_paths
-
-
-def test_cache_constructor_with_existing_evaluations(tmp_path):
-    cache_dir = tmp_path / "cache"
-    evaluations_dir = cache_dir / "evaluations"
-    evaluations_dir.mkdir(parents=True)
-    evaluation_paths = [
-        evaluations_dir / "evaluation-1000.json",
-        evaluations_dir / "evaluation-999.json",
-        evaluations_dir / "evaluation-900.json",
-        evaluations_dir / "evaluation-889.json",
-        evaluations_dir / "evaluation-700.json",
-        evaluations_dir / "evaluation-88.json",
-        evaluations_dir / "evaluation-70.json",
-    ]
-    for evaluation_file in evaluation_paths:
-        evaluation_file.touch()
-
-    cache = Cache(cache_dir)
-
-    assert cache.cache_root_directory == cache_dir
-    assert cache_dir.exists()
-    assert cache.evaluations_dir == cache_dir / "evaluations"
-    assert cache.evaluations_dir.exists()
-    assert cache.all_evaluation_paths == evaluation_paths
-    assert cache.recent_evaluation_path == evaluation_paths[0]
-    for i, evaluation_path in enumerate(evaluation_paths):
-        assert cache.evaluation_path(i) == evaluation_path
-
-
 def test_save_evaluation(tmp_path, mock_time):
     time = 12300566
     cache_dir = tmp_path / "cache"
@@ -142,7 +69,7 @@ def test_cache_save_evaluation_fails_when_no_root_dir_was_set():
 
 
 @parametrize(argnames="invalid_evaluation_index", argvalues=[-1, 10])
-def test_cache_get_evaluation_with_invalid_context_(tmp_path, invalid_evaluation_index):
+def test_cache_get_evaluation_with_invalid_index(tmp_path, invalid_evaluation_index):
     cache_dir = tmp_path / "cache"
     evaluations_dir = cache_dir / "evaluations"
     evaluations_dir.mkdir(parents=True)
